third_party/WebKit/LayoutTests/platform/win/editing/selection/modify_extend/extend_by_character-expected.txt

This is a testharness.js-based test.
PASS extend forward character on characters
PASS extend forward character through image
FAIL extend forward character through multiple spaces LayoutTests/editing/selection/modify_extend/extend_by_character.html:42:12)
	 expected <div contenteditable> <i>F    ^and seven</i> years <b>  as </b>our fathers  f upon this continent, a new nation, conceived    in Liberty,   and dedicated to the proposition that all  <br>men are created equal|. </div>,
	 but got  <div contenteditable> <i>F    ^and seven</i> years <b>  as </b>our fathers  f upon this continent, a new nation, conceived    in Liberty,   and dedicated to the proposition that all  <br>men are created equ|al. </div>,
	 sameupto <div contenteditable> <i>F    ^and seven</i> years <b>  as </b>our fathers  f upon this continent, a new nation, conceived    in Liberty,   and dedicated to the proposition that all  <br>men are created equ
PASS extend backward character through image
FAIL extend backward character through multiple spaces LayoutTests/editing/selection/modify_extend/extend_by_character.html:92:12)
	 expected <div contenteditable> <i>|F    and seven</i> years <b>  as </b>our fathers  f upon this continent, a new nation, conceived    in Liberty,   and dedicated to the proposition that all  <br>men are created equ^al. </div>,
	 but got  <div contenteditable> <i>F    |and seven</i> years <b>  as </b>our fathers  f upon this continent, a new nation, conceived    in Liberty,   and dedicated to the proposition that all  <br>men are created equ^al. </div>,
	 sameupto <div contenteditable> <i>
PASS extend forward character over element
Harness: the test ran to completion.
